requirements-dev.txt
codecov>=2.0.5
coverage>=4.2
cython>=0.25.1
flake8>=3.0.4
path.py>=8.2.1
pytest>=3.0.3
pytest-cov>=2.4.0
pytest-mock>=1.4.0
pytest-runner>=2.9
pytest-virtualenv>=1.2.5
six>=1.10.0
